+++ tsdb/engine/tsm1/engine.go
@@ -61,9 +61,11 @@ type Engine struct {
 	wg                 sync.WaitGroup
 	compactionsEnabled bool
 
-	path      string
-	logger    *log.Logger
-	logOutput io.Writer
+	path         string
+	logger       *log.Logger // Logger to be used for important messages
+	traceLogger  *log.Logger // Logger to be used when trace-logging is on.
+	logOutput    io.Writer   // Writer to be logger and traceLogger if active.
+	traceLogging bool
 
 	// TODO(benbjohnson): Index needs to be moved entirely into engine.
 	index             *tsdb.DatabaseIndex
@@ -95,11 +97,7 @@ type Engine struct {
 // NewEngine returns a new instance of Engine.
 func NewEngine(path string, walPath string, opt tsdb.EngineOptions) tsdb.Engine {
 	w := NewWAL(walPath)
-	w.LoggingEnabled = opt.Config.WALLoggingEnabled
-
 	fs := NewFileStore(path)
-	fs.traceLogging = opt.Config.DataLoggingEnabled
-
 	cache := NewCache(uint64(opt.Config.CacheMaxMemorySize), path)
 
 	c := &Compactor{
@@ -108,7 +106,12 @@ func NewEngine(path string, walPath string, opt tsdb.EngineOptions) tsdb.Engine 
 	}
 
 	e := &Engine{
-		path:              path,
+		path:         path,
+		logger:       log.New(os.Stderr, "[tsm1] ", log.LstdFlags),
+		traceLogger:  log.New(ioutil.Discard, "[tsm1] ", log.LstdFlags),
+		logOutput:    os.Stderr,
+		traceLogging: opt.Config.TraceLoggingEnabled,
+
 		measurementFields: make(map[string]*tsdb.MeasurementFields),
 
 		WAL:   w,
@@ -127,7 +130,12 @@ func NewEngine(path string, walPath string, opt tsdb.EngineOptions) tsdb.Engine 
 		enableCompactionsOnOpen:       true,
 		stats: &EngineStatistics{},
 	}
-	e.SetLogOutput(os.Stderr)
+
+	if e.traceLogging {
+		e.traceLogger.SetOutput(e.logOutput)
+		fs.enableTraceLogging(true)
+		w.enableTraceLogging(true)
+	}
 
 	return e
 }
@@ -302,17 +310,28 @@ func (e *Engine) Close() error {
 	return e.WAL.Close()
 }
 
-// SetLogOutput sets the logger used for all messages. It must not be called
-// after the Open method has been called.
+// SetLogOutput sets the logger used for all messages. It is safe for concurrent
+// use.
 func (e *Engine) SetLogOutput(w io.Writer) {
-	e.logger = log.New(w, "[tsm1] ", log.LstdFlags)
+	e.logger.SetOutput(w)
+
+	// Set the trace logger's output only if trace logging is enabled.
+	if e.traceLogging {
+		e.traceLogger.SetOutput(w)
+	}
+
 	e.WAL.SetLogOutput(w)
 	e.FileStore.SetLogOutput(w)
+
+	e.mu.Lock()
 	e.logOutput = w
+	e.mu.Unlock()
 }
 
 // LoadMetadataIndex loads the shard metadata into memory.
 func (e *Engine) LoadMetadataIndex(shardID uint64, index *tsdb.DatabaseIndex) error {
+	now := time.Now()
+
 	// Save reference to index for iterator creation.
 	e.index = index
 
@@ -347,6 +366,7 @@ func (e *Engine) LoadMetadataIndex(shardID uint64, index *tsdb.DatabaseIndex) er
 		}
 	}
 
+	e.traceLogger.Printf("Meta data index for shard %d loaded in %v", shardID, time.Since(now))
 	return nil
 }
 
@@ -674,6 +694,7 @@ func (e *Engine) WriteSnapshot() error {
 	defer func() {
 		if started != nil {
 			e.Cache.UpdateCompactTime(time.Now().Sub(*started))
+			e.logger.Printf("Snapshot for path %s written in %v", e.path, time.Since(*started))
 		}
 	}()
 
@@ -708,7 +729,9 @@ func (e *Engine) WriteSnapshot() error {
 	// The snapshotted cache may have duplicate points and unsorted data.  We need to deduplicate
 	// it before writing the snapshot.  This can be very expensive so it's done while we are not
 	// holding the engine write lock.
+	dedup := time.Now()
 	snapshot.Deduplicate()
+	e.traceLogger.Printf("Snapshot for path %s deduplicated in %v", e.path, time.Since(dedup))
 
 	return e.writeSnapshotAndCommit(closedFiles, snapshot)
 }
@@ -772,6 +795,7 @@ func (e *Engine) compactCache() {
 			e.Cache.UpdateAge()
 			if e.ShouldCompactCache(e.WAL.LastWriteTime()) {
 				start := time.Now()
+				e.traceLogger.Printf("Compacting cache for %s", e.path)
 				err := e.WriteSnapshot()
 				if err != nil && err != errCompactionsDisabled {
 					e.logger.Printf("error writing snapshot: %v", err)
@@ -970,6 +994,7 @@ func (e *Engine) compactTSMFull() {
 
 // reloadCache reads the WAL segment files and loads them into the cache.
 func (e *Engine) reloadCache() error {
+	now := time.Now()
 	files, err := segmentFileNames(e.WAL.Path())
 	if err != nil {
 		return err
@@ -989,6 +1014,7 @@ func (e *Engine) reloadCache() error {
 		return err
 	}
 
+	e.traceLogger.Printf("Reloaded WAL cache %s in %v", e.WAL.Path(), time.Since(now))
 	return nil
 }